fix:Homepage moved to pages/Home, created file Home.elements.js
